Updated proposal loading to avoid sorting (#363)
Summary:
Updated the code and docstring for loading precomputed proposals. The docstring now mentions the format of the proposal dict so that it is easier for users to add their own region proposal files. The loading code no longer requires the proposal dict be sorted in the same order as the dataset.
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/363

Differential Revision: D18676355

Pulled By: ppwwyyxx

fbshipit-source-id: 2b17d8d3f8d18be07037faac323b01434534417e