Fix multiple processes writing coco json to same file (#543)
Summary:
When CocoEvaluator with multiple workers using a custom loader, it appears that each worker converts the Detectron2 dataset to COCO json and then dump to disk. This can cause the output file to be corrupted. Here my fix is to only convert/dump on the main process/
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/543

Differential Revision: D19164257

Pulled By: ppwwyyxx

fbshipit-source-id: 4d6be0bb3e0520f55e908ffa1652c1ad3d55da59