Make application of norm weight decay more robust
Summary: The current logic for finding feature normalization modules is based on fragile string matching. For models that include norm modules in, for example, an `nn.Sequential` the logic fails to detect the module (because it appears like `0.weight`, `0.bias`, ...). This diff makes the matching more robust by matching against a set of normalization module types. The downside is that this set needs to be maintained if new normalization types, that do not inherit from any of these, are added.

Reviewed By: ppwwyyxx

Differential Revision: D19393513

fbshipit-source-id: e81fe36df7e09ec33addbca515cefdfd86c69fb9