Add typing to ROIHeads
Summary:
I noticed an error in the docs, which pushed me to add typing to these functions. Since we're already doing the work of documenting all of the types, we should have the type checker verify they're correct.

Ideally we can switch to something like https://github.com/agronholm/sphinx-autodoc-typehints for generating the docs and remove the types from the docstrings.

Reviewed By: ppwwyyxx

Differential Revision: D19814927

fbshipit-source-id: cfde9ccf02d197b86f242d49bbdc5b2ede4608a6