ROIHeads: pass self.input_shape instead of saving channels & strides
Summary:
The saved channels & strides doesn't provide much benefits: it's similar amount of code to just use
the original input shape given.

Having access to the input shape will allow passing it to heads, for heads that do take all
features, or if we want to support different heads that have their own "IN_FEATURES".
This will be useful in the following diffs.

Reviewed By: rbgirshick

Differential Revision: D19852188

fbshipit-source-id: 1657952161ba7dc490404bf99b9c2c7bfcacb7fd