fix inference tests
Summary:
The existing tests are broken because the model loaded in the tests does not correspond to the config (the config segmentation layer has 2 channels, the model has 15 channels).

This diff performs inference with a model that corresponds to the config.
It also adjusts test scripts to use 2 images per GPU.

Reviewed By: patricklabatut

Differential Revision: D20765951

fbshipit-source-id: 0a695e5e5b1847f1ddd72d091ba284ade0616084