Adding histograms to tensorboard
Summary:
Hey,

as this is a small PR and I already implemented this feature for myself I did not open an issue.

Description:
I changed EventStorage and TensorboardXWriter such that histograms (for monitoring weights or gradients) can be saved to tensorboard.

To make this implementation as efficient as possible I calculate the histogram statistics for the add_histogram_raw function of the PyTorch SummaryWriter (instead of saving weight matrices). In general, I tried to match the information flow to that of the put_image function.

In my personal implementation I extended the EventStorage and TensorboardXWriter, but I thought that this might be a feature that Detectron could offer out of the box.

A possible example usage could be a Hook:
```
class HistogramHook(HookBase):
    def after_step(self):
        storage = self.trainer.storage
        hist_period = self.trainer.cfg.HIST_PERIOD

        if hist_period > 0 and storage.iter % hist_period == 0:
            self.add_histogram(self.trainer.model, storage)

    def add_histogram(self, model, storage):
        for i, (tag, value) in enumerate(model.named_parameters()):
            # E.g. add histograms for parameters of the Mask R-CNN mask head
            if tag.split(".")[1] == "mask_head":
                storage.put_histogram(tag, value.data.cpu().numpy())
```
Feedback is very much appreciated! :)

Best,
David
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/1265

Reviewed By: rbgirshick

Differential Revision: D21223968

Pulled By: ppwwyyxx

fbshipit-source-id: 92a7b3d04fd1b45b4fd35fb9a814e529ba40ab9d