Correctly apply EXIF orientation to images
Summary:
Workaround for the PIL bug https://github.com/python-pillow/Pillow/issues/3973 - the exif orientation metadata is used to correctly transpose a PIL image when reading an image. The current approach silences any exceptions here, leading to strange behavior, such as issue https://github.com/facebookresearch/detectron2/issues/933. This is indeed a PIL issue, but the fix here performs the same function that is necessary to correctly orient images. The only differences between the committed function `apply_exif_orientation`  and the PIL `exif_transpose` is that a copy of the image is not generated if no transpose is needed, and that the exif data is not updated in the transposed image. As `read_image` returns a NumPy ndarray, this information is discarded, anyway.
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/1666

Differential Revision: D22238978

Pulled By: ppwwyyxx

fbshipit-source-id: 497c97b99fbfcf9857cbdf7516df10da4b42278d