Fix numerical instability of rotated IoU (issue 2154 & issue 2167)
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/2178

https://github.com/facebookresearch/detectron2/issues/2154

- Fixed numerical instability of rotated IoU and added two more unit tests covering the two examples reported.

When computing the intersection area, it doesn't hurt if we have more (duplicated/approximate) intersections/vertices than needed, while it can cause drastic difference if we miss an intersection/vertex. Therefore, we add an epsilon to relax the comparisons between the float point numbers that decide the intersection points.

- Fixed unit tests coverage by changing
```
["cpu"] + ["cuda"] if torch.cuda.is_available() else []
```
to
```
["cpu"] + (["cuda"] if torch.cuda.is_available() else [])
```
since otherwise the result would be [] instead of ["cpu"] on cpu machines.

Reviewed By: ppwwyyxx

Differential Revision: D24498800

fbshipit-source-id: f4a04619e9da926d3d3d6291482aa246de6965db