Replace copy_ op in Box2BoxTransform.apply_deltas
Summary:
The goal of this PR is to remove dead-end branches from the graph caused by in-place operators (such as `copy_`). The improvement in the graph structure is needed to make it easier to use the model with 3rd party inference frameworks (e.g. TVM, ONNX, etc).

Related discussions:
- https://github.com/facebookresearch/detectron2/discussions/2678

ppwwyyxx

Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/2685

Reviewed By: theschnitz

Differential Revision: D26745044

Pulled By: ppwwyyxx

fbshipit-source-id: 7e25ab3a7a19812581f36a585cde2d2493489090