let PointRend inference start with a mask computed by running point_head on regular grid
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/2762

Before, PointRend inference starts with coarse mask. Now we let it start by running point_head on a
regular grid.

This only changes the behavior of the model when `SUBDIVISION_NUM_POINTS <= 4*INIT_RESO**2=196`, which is very rare.
Even when it does change the behavior, it doesn't seem to decrease AP.

Differential Revision: D27149708

fbshipit-source-id: 2b74f556d5228e18dafcc05541dc3b16f4ab8319