add keyframe helper to video keyframe dataset
Summary:
Adds a possibility to specify keyframes for videos in `VideoKeyframeDataset`. This can be useful to accelerate data loading:
1. If video keyframes are provided, this helps avoid `list_keyframes` operation;
2. If video keyframes with detections are provided, this avoids `list_keyframes` operation and increases detection rate in `InferenceBasedLoader`, which is based on `VideoKeyframeDataset`.

Keyframes are specified in CSV format, e.g.:
```
video_id,keyframes
0,[]
1,"[1,11,21]"
2,[1]
...
```
If a given `video_id` is not present in the helper data, the usual procedure is performed (`list_keyframes` + `read_keyframes`), which requires 2 passes over video keyframes: 1 demuxing + 1 decoding. Otherwise, the list of keyframes is taken from the list specified in the helper data, which avoids 1 demuxing pass over the video.

Reviewed By: patricklabatut

Differential Revision: D27136885

fbshipit-source-id: 2a2099d9231970d5a91d4e0cd8472856b57ba989