perform training logic in train_net.py
Summary:
The new system no longer has pre-defined structure, so using a DefaultTrainer(cfg) makes it confusing what cfg keys are needed.

The code in DefaultTrainer is short enough to live in train_net so users can see directly what config is needed. This way we also avoid reinventing a trainer that takes new configs. We'll be able to remove newconfig/defaults.py then (or keep it private for reference)

Reviewed By: theschnitz, alexander-kirillov

Differential Revision: D27716057

fbshipit-source-id: 248446d19bfaf601653bb25bb09765f99e3a8432