Add animal category to bootstrapping datasets
Summary:
In the bootstrapping datasets, we might not have annotated images, but we often know very well to which animal categories they belong. This diff aims to pass down this information.

For the details of the Instagram Datasets and `build_fb`, check the following diff D27045763

Category to class mapping should look like that and aims to convert "layman's terms" categories, used in the datasets, to category classes (and then to mesh names with CLASS_TO_MESH_NAME_MAPPING)

{F506421983}

Reviewed By: vkhalidov

Differential Revision: D27045669

fbshipit-source-id: 3865c36e622813b7591a3a969a2fa58295a48744