Make vis module importable, remove required torch and torchvision
Summary:
In PR https://github.com/facebookresearch/detectron2/issues/3311 we made densepose a pip-installable Python package. However, we forgot to include an `__init__.py` so that the `vis` module is importable. This PR adds that `__init__.py`.

It was also mentioned in that PR that there is a preference not to require torch and torchvision requirements versions:
https://github.com/facebookresearch/detectron2/pull/3311#issuecomment-892513481

This PR also removes those requirements from the `setup.py` and lists them instead in the prerequisites in the documentation.

Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/3382

Reviewed By: patricklabatut

Differential Revision: D30462761

Pulled By: vkhalidov

fbshipit-source-id: e58985ff24cc75ff1d7905a863fa040e3be99bc3