fix rotated bounding box tracing inference error
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/4249

Fixes run time error:
RuntimeError: Output 0 of SelectBackward0 is a view and is being modified inplace. This view is the output of a function that returns multiple views. Such functions do not allow the output views to be modified inplace. You should replace the inplace operation by an out-of-place one.

Reviewed By: zhanghang1989

Differential Revision: D36433376

fbshipit-source-id: e337e63e2ce09632d42db79a7b6e7b050ceb325c