Add assertion error if model not c2 compatible
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/4248

The caffe 2 model exported was not handling MODEL.RPN.BBOX_REG_WEIGHTS) different of (1., 1., 1., 1.) or (1., 1., 1., 1., 1.) correctly because the config is not passed to caffe2 generate_proposals config. Modifying it won't slove the issue as there will be an api change customers like boltnn will have to manage. This fix is to fail during caffe2 export for non supported config to avoid non regression behavior troubleshooting

Reviewed By: sstsai-adl

Differential Revision: D36489096

fbshipit-source-id: c955684aa64d35905477db2390c019bf5bc0fbd8