Add support for ONNX-only (original PR #4120)
* add_export_config API is publicly available exposed even when caffe2 is not
compiled along with PyTorch (the new default behavior on latest PyTorch).
A warning message informing users about its deprecation on future versions is also added

* `tensor.shape[0]` replaces `len(tensor)` and `for idx, img in enumerate(tensors)`
replaces `for tmp_var1, tmp_var2 in zip(tensors, batched_imgs)`
so that the tracer does not lose reference to the user input on the graphs.

* Added unit tests (tests/torch_export_onnx.py) for detectron2 models

* ONNX is added as dependency for the CI to be able to run the aforementioned tests

* Added custom symbolic functions to allow CI pipelines to succeed.
The symbolics are needed because PyTorch 1.8, 1.9 and 1.10 adopted
by detectron2 have several bugs. They can be removed when 1.11+ is
adopted by detectron2's CI infra