Add pyproject.toml for PEP-518 compliance The new pyproject.toml specifies the minimum requirements necessary to build the project's wheel. This makes this project up to the PEP-518 standard.
Specifically, this change adds torch as a build-time dependency. The
version of torch can be set by the installer when they specify a version
of torch in their own project settings (e.g. a pyproject.toml,
requirements.txt, setup.py, environment.yaml, etc.). Alteratively,
if torch is in the current installing environment, that version will be
used. If not specified, the latest torch version will be used.

This change preserves using  for project building.

python-poetry/poetry#2113