make the offload function customizable for DatasetFromList
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/4626

Previously we use `serialize: bool` to control if we want to offload the `DatasetFromList` storage to numpy. This diff generalize the "serialize" to "offload", and make the "offload function" customizable so that we can switch between implementations.

The setting of `offload function` is done by context manager in order to avoid passing this argument all the way down.

Reviewed By: sstsai-adl

Differential Revision: D40818736

fbshipit-source-id: ed1b47eea86546def6c06f78bc12d6edf267df28