Convert mock_torch_nn_functional_interpolate to decorator
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/4820

We're working on onboarding Stinson models onto Executorch, the next generation runtime. In order to do so we must first export models by tracing it with Torch Dynamo. However, as Dynamo is actively being developed, it doesn't handle all features, such as tracing generators, which was encountered when exporting the Keyboard Tracking model.

However, it looks like the `mock_torch_nn_functional_interpolate` function might also work as a decorator rather than
a generator, which dynamo is able to trace through.

Differential Revision: https://internalfb.com/D43400473

fbshipit-source-id: 518df3e954c2241e6e504bd10d88ba8340944be5