Fix data dependent flow
Summary:
Dynamo is unable to handle the data dependent flow that is using values within a tensor as the shape to create another tensor. It is unable to infer the value of the tensor since that value could be different depending on inputs.

However, looking at the code, it seems this might not be needed. This is an implementation detail of [torch.ops._caffe2.BoxWithNMSLimit](https://fburl.com/code/mxmrwqv5), but the value inside nms_outputs[4] is just the size of `roi_batch_splits` which is a single integer. So, we can just use the size of `roi_batch_splits` to create the batch ids

Differential Revision: D43641703

fbshipit-source-id: b3c0ecc733d983bdf9a8b0440f53a76de2570a61