Lazy initialize tensorboard
Summary:
Every time a summary writer object is created it'll end up creating a file in the passed in logdir with just metadata of 40 bytes. This will cause a lot of spam with every rank initializing the file aggressively (I've done this myself when adding support for Adhoc logging to PyPer to a lot of angry complaints).

Report at https://fb.workplace.com/groups/723537759122220/posts/745046770304652/

Differential Revision: D43863071

fbshipit-source-id: ff36ab3b6bde185f41d71b5e4ddf97e671dff6ca