Add an option to zero out the gradient before the forward
Summary:
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/4905

Currently the optimizer zeros the gradients after the forward and before the backward. In a recent PyTorch change, it set all gradients to None by default. This has a benefit of reducing the memory consumption (since all gradients are None).

However, doing this after the forward does not provide any memory saving, since the the memory consumption is maximum at the end of forward.

It doesn't matter whether the gradient is set to None before the forward or after the forward. So we should set it before the forward the enjoy the memory saving.

We add a flag to enable it instead of doing it by default for now. Since people can override the zero_grad function (as the comment indicates), we do not know exactly what is done inside the function. This is to be on the safe side so the current flows may not be broken.

Once we have gone through the existing flows, we should make the flag enabled by default.

Reviewed By: tglik

Differential Revision: D44264848

fbshipit-source-id: 9d79eef6d667ebc9753d3ac22573d42f72f98622