Revert D44153511: Multisect successfully blamed D44153511 for test or build failures
Summary:
This diff is reverting D44153511
D44153511: [D2GO] TensorBoard changes by hamzaj has been identified to be causing the following test or build failures:

Tests affected:
- [mobile-vision/d2go/tests:init_tests - test_train_net_main (d2go.tests.fb.test_tools_train_net.TestToolsTrainNet)](https://www.internalfb.com/intern/test/844425016192052/)

Here's the Multisect link:
https://www.internalfb.com/multisect/1886535
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: miqueljubert

Differential Revision: D44995942

fbshipit-source-id: 45fab0356724417fe1dcdfdee6336ee5bf194b1b