Fix unused vars in _decode_multi_level_predictions
Summary:
This is a simple PR that fixes two unused arguments in `_decode_multi_level_predictions` which may lead to confusion:
- `score_thresh`
- `topk_candidates`

https://github.com/facebookresearch/detectron2/blob/2254a10de3a06536fb1d18b55417cd5ece192f59/detectron2/modeling/meta_arch/dense_detector.py#L235-L255

This hasn't been an issue because, luckily, both concrete `DenseDetector`s available in detectron2 (FCOS and RetinaNet) pass the `self.test_*` as arguments when calling it. This is misleading, nonetheless.

An alternative approach would be to remove these arguments, but that would breaking change.

Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/4993

Reviewed By: jiaxuzhu92

Differential Revision: D46562843

fbshipit-source-id: e90c569c2f4d26fcb263265c6d5e1014fbf0cbf8