replace np.int with int [1]
Summary:
`numpy.int` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `int` and should be safe to replace with `int`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.int\b' | xargs perl -pi -e 's,\bnp\.int\b,int,g'
```

Differential Revision: D46607065

fbshipit-source-id: ab4b835567c7c9540e54ab6c73a085ee8b0c48ee