Support training models on Smart Glasses YUV color data.
Summary:
The existing "YUV-BT.601" format seems to be a misnomer(is actually BT.470), so adding a new format and conversion methods for Smart Glasses Camera YUV inputs, the actual BT.601.

The only caveat here is that these models will expect full resolution U and V channels. Upscaling needs to happen outside the model or in a torchscript wrapper - TBD.

Reviewed By: srihari2761

Differential Revision: D48282628

fbshipit-source-id: b1a5d0c30dc338c84f30bcca62130ab4b3f58511