Find all invalid dataset names and throw KeyError once.
Summary:
Currently, it is painful to validate that a list of dataset names are in the DatasetCatalog, since each invalid dataset name causes a new KeyError to be thrown.

Instead, this diff checks all the dataset names at once, throwing a KeyError if any dataset is invalid and listing all invalid dataset names.

Reviewed By: wat3rBro

Differential Revision:
D49252165

Privacy Context Container: L1211256

fbshipit-source-id: 00fac3510fbb75c1cda72530c8d366b3027bb5ab