add DATALOADER.REPEAT_SQRT
Summary:

X-link: https://github.com/fairinternal/detectron2/pull/602

For sampler **RepeatFactorTrainingSampler**, current per-category weight is computed as **1/sqrt(frequency)**.

This works fine on LVIS but is not sufficient in highly imbalanced data we have for person segmentation.

Thus we add an argument **DATALOADER.REPEAT_SQRT**. If false, we compute per-category weight as **1/frequency**.

This is change is entirely back-compatible.

Differential Revision: D55355021