Address Breaking Detectron Test
Summary: Moved evaluator builder function to eval callback function to remove cfg as a parameter to RotatedCocoEvaluator as it is of type DictConfig and not CfgNode. As a result, cfg.TEST.KEYPOINT_OKS_SIGMAS field is unused for RotatedCocoEvaluator in config file (should be okay unless want to set kpt_oks_sigmas through config file).

Reviewed By: luyich

Differential Revision: D66061894

fbshipit-source-id: 68684e661b69bc7e1458d8775d5766356b865e45